폴더 구조 변경 및 테스트
