+++ src/main/java/com/service/RSIranking/dto/KospiSecuritiesStockDto.java
@@ -21,7 +21,7 @@ public class KospiSecuritiesStockDto implements Serializable, StockDto {
     private String mktNm;
     private String type;
     private Boolean isPublicStock;
-    private boolean checked = false;
+    private boolean checked;
 
     /**
      * JSON 데이터 SecuritiesStockDto로 변환
